Update s7-enumerate.nse
Changed spacing on some of the lines from copy/paste to be standard. Also changed an if else statement inside of the parse_response function to make it more simple. Added a second function to parse the second result, this was done to remove the static ONE and TWO for the code variable that was being passed in.