Set version info to Nmap-style values, clean up.
Nmap's port.version.name field should be a service name, not a product
name. Since S7 runs over the TPKT protocol which is a port of ISO-TSAP
to TCP, we keep the iso-tsap service name. The version information
doesn't seem relevant here, since it is reported in the script output,
so the product is simply "Siemens S7 PLC" without version.

Removed function tablelength, since stdnse.output_table can be checked
for number of keys using the length operator (#). Not usually possible,
but output_tables are weird.